requirements.txt
-e git://github.com/sehmaschine/django-grappelli.git#egg=django-grappelli
-e git://github.com/madisona/django-google-maps.git#egg=django-google-maps

#PIL
sorl-thumbnail
django-hosts
django-social-auth

